audiobook_service/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobook_service"
version = "0.1.0"
description = "Audiobook generator service using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai-agents[voice]",
    "python-docx",
    "PyPDF2",
    "python-multipart",
]